BC: Make the routes config dynamically
In the `config/laravel-impersonate.php` the `take_redirect_to` and `leave_redirect_to` are hard coded urls. This is fine as long as you don't construct your routes dynamically.

In our case we hook into the `RouteServiceProvider` to add a language prefix `de` or `en`, depending on the user settings.

Since `route()` in a config file leads to a "Route not defined" error, because Laravel loads the configuration before the routes are registered, the only way to tackle this is to expect a route name.

The current implemenation is a BC break. A better implementation would could detect a url and a route and decide this on its own. This is just a first draft to see if you are interested in such a change.